refactor: Improving the makefile
In the last refactoring, I transfer several assignments from main.cpp to
the FractalCreator class. However, I forgot to consider that this new
class is the one that contains the header of several others. I changed
the order of calling some targets so that the FractalCreator object can
be created regardless of main.cpp.